binder finish

+++ src/execution/filter_executer.hpp
@@ -1,15 +1,18 @@
 #pragma once
 
 #include "execution/execution.hpp"
+#include "sql/parser/ast.hpp"
 
 class FilterExecuter : public Executer {
 public:
-  FilterExecuter(Expression* expr,Executer* child) {
+  FilterExecuter(Expression *expr, Executer *child)
+  {
     expression_ = std::unique_ptr<Expression>{expr};
     children_.emplace_back(child);
   }
 
-  RC next(Tuple* t) override {
+  RC next(Tuple *t) override
+  {
     RC rc = RC::SUCCESS;
     for (;;) {
       rc = children_[0]->next(t);
@@ -18,7 +21,7 @@ public:
       }
 
       auto ok_value = expression_->evaluate(t, children_[0]->schema_.get());
-      if (ok_value.get<int>() == 1) {
+      if (ok_value.get<long>() == 1) {
         break;
       }
     }
